backend/requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
python-jose
passlib[bcrypt]
sqlalchemy
psycopg2-binary
requests
phonenumbers
python-multipart
pydantic[email]
opencv-python
numpy
httpx
fastapi-limiter
alembic
